Refactoring Date Picker
